requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
# Machine Learning & Noise Detection
scikit-learn>=1.3.0
cleanlab>=2.5.0
# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
# Optional Dashboard/Web Interface
streamlit>=1.24.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0
# Development & Testing
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
scipy>=1.11.0
# For data examples
scikit-learn-datasets>=0.1.0
kagglehub>=0.1.0
